Move test_cli.py out of redundant pms subpackage
All the other tests are targetting code in the pms package so this
"pms" subpackage containing a single test file was confusing to me.

Since the tests are primarily targetting code in the pms.cli package
it makes sense to move the file up one level so it roughly matches
the directory structure of the code under test. Having all the tests
follow a consistent structure will make it easier to add new ones.

Note: the existing tests are a mix of unit and integration isolations;
this change just makes the existing structure a bit clearer without
trying to introduce further separation of the tests by isolation level.