Add tests for SensorReader context manager
This uses a mock_serial library to test the context manager makes
the expected read/write calls to the underlying serial device. By
specifying the sensor as "PMSx003" we can hard-code the necessary
stubs so that the mock device interacts properly with the reader.

While other tests cover parts of the reader [^1], much of the code
is mocked using the "capture" fixture [^2]. This test adds some of
the missing coverage before we change the class in later commits.

Note that the tests can be brittle due to use of "sys.exit" in the
reader code: if something goes wrong and the conditional branches
with "sys.exit" run, pytest will terminate early and incorrectly
show tests as passing. We'll try and improve this in a later commit.

[^1]: https://github.com/benthorner/PyPMS/blob/d6e6333d861c6780ba2d232cc8d530d8c6693f39/tests/test_cli.py
[^2]: https://github.com/benthorner/PyPMS/blob/d6e6333d861c6780ba2d232cc8d530d8c6693f39/tests/conftest.py#L110-L166