Raise exceptions instead of calling sys.exit
This is a step towards making the code in the reader module work in
a library context, as well as making it possible to test the happy
and unhappy paths reliably - without sys.exit interfering.

In order to keep the same CLI behaviour - and not print a stacktrace
- each command that calls the reader now needs to handle the error
and call sys.exit(1) itself. Although it's duplicated, exiting the
program at the top-level makes it easier to reason about.