Support getting individual readings from readers
This is part of adapting the reader module to support its use as a
library. Previously it was only possible to read *all* the data in
one go, which meant that error handling was hard to customise.

Because the __call__ method needs both the "buffer" and the decoded
"ObsData" the new "read_one" method need to return both together.
This is achieved using a new "Reading" type to hold both values.

Introducing "Reading" isn't ideal as we now have three types that are
used to hold data. In future we could look at breaking changes to
converge on a single type for sensor data.