DRY-up reader sample / sleep into utility class
This makes the main control flow of the __call__ method clearer
for each reader by extracting common housekeeping code to a new
Sampler utility class. The sampling behaviour is the same between
the readers, except:

- SensorReader has an optional sleep between readings. To emulate
this outside of a generator context, the new Sampler class stores
the last reading it sampled.

- MessageReader only allows a single sample countdown. To emulate
this the sampler is instantiated at the object level, rather than
being created with each __call__.

A quirk of the original sample behaviour was also that a negative
number of samples would still result in a single sample, which is
a reasonable alternative to rejecting the (arguably) invalid value.
This preserves that.