Make sensor pre-heat overridable in SensorReader
This is another part of making the reader class more amenable for
use in a library, allowing the library to control sleep behaviour.

For example, in an event loop system, it makes more sense to check
if "pre_heat_seconds" has elapsed and then give control back to the
event loop - as opposed to blocking the whole system for X seconds.